+++ ebi/ols/api/base.py
@@ -179,11 +179,7 @@ class DetailClientMixin(BaseClient):
                 if not unique:
                     return elms
                 else:
-                    for elem in elms:
-                        if elem.is_defining_ontology:
-                            return elem
-                    logger.warning('Unable to fin item %s defined in an ontology', logger_id)
-                return None
+                    return next((x for x in elms if x.is_defining_ontology), elms[0])
             return self.elem_class_instance(**document.data)
         except coreapi.exceptions.ErrorMessage as e:
             if 'status' in e.error:
@@ -371,7 +367,6 @@ class ListClientMixin(BaseClient):
         elif isinstance(item, int):
             page = item // self.page_size
             index = item % self.page_size
-            logger.debug('Current page %s. %s translated to %s in page %s', self.page, item, index, page)
             if page != self.page:
                 self.fetch_page(page)
                 self.index = index